Consolidate methods of getting subclasses
