+++ src/Post/Loader.php
@@ -20,7 +20,7 @@ use Taco\Util\Str as Str;
 global $taxonomies_infos;
 $taxonomies_infos = array();
 
-class Loader
+class Loader extends \Taco\BaseLoader
 {
         
     /**
@@ -143,22 +143,4 @@ class Loader
         return $count;
     }
     
-    
-    /**
-     * Get all subclasses
-     * @return array
-     */
-    public static function getSubclasses()
-    {
-        $subclasses = array();
-        foreach (get_declared_classes() as $class) {
-            if (method_exists($class, 'isLoadable') && $class::isLoadable() === false) {
-                continue;
-            }
-            if (is_subclass_of($class, 'Taco\Post')) {
-                $subclasses[] = $class;
-            }
-        }
-        return $subclasses;
-    }
 }
